@@ -1,11 +1,12 @@
 #ifndef __BIG_INT_H__
 #define __BIG_INT_H__
 
+#include "utils.h"
+
+#include <cstdint>
+#include <math.h>
 #include <vector>
 #include <string>
-#include <limits>
-#include <math.h>
-
 #include <type_traits>
 
 #define LOGGING(a)
@@ -18,21 +19,129 @@ struct DivModResult { D d; R r; };
 class IBigIntLike { };
 
 template <typename T>
-using isBigIntLike = typename std::enable_if<std::is_base_of<IBigIntLike, T>::value, T>::type;
+concept is_BigInt_like1 = std::is_base_of_v<IBigIntLike, T>;
 
-template <typename T, typename Tt>
-using isBigIntLikeT1 = typename std::enable_if<std::is_base_of<IBigIntLike, T>::value, Tt>::type;
+template <typename T1, typename T2>
+concept is_BigInt_like2 = std::is_base_of_v<IBigIntLike, T1> && std::is_base_of_v<IBigIntLike, T2>;
 
-template <typename T1, typename T2, typename Tt>
-using isBigIntLikeT2 = typename std::enable_if<std::is_base_of<IBigIntLike, T1>::value && std::is_base_of<IBigIntLike, T2>::value, Tt>::type;
+template <typename T1, typename T2, typename T3>
+concept is_BigInt_like3 = std::is_base_of_v<IBigIntLike, T1> && std::is_base_of_v<IBigIntLike, T2> && std::is_base_of_v<IBigIntLike, T3>;
+
+
+class BigInt;
+
+CONSTEXPR_AUTO to_string_base2(const BigInt &v) -> std::string;
+CONSTEXPR_AUTO to_string_base8(const BigInt &v) -> std::string;
+CONSTEXPR_AUTO to_string_base10(const BigInt &v) -> std::string;
+CONSTEXPR_AUTO to_string_base16(const BigInt &v) -> std::string;
+CONSTEXPR_AUTO to_string(const BigInt &v) -> std::string;
+
+CONSTEXPR_AUTO from_string_base2(const std::string &input) -> BigInt;
+CONSTEXPR_AUTO from_string_base8(const std::string &input) -> BigInt;
+CONSTEXPR_AUTO from_string_base10(const std::string &input) -> BigInt;
+CONSTEXPR_AUTO from_string_base16(const std::string &input) -> BigInt;
+CONSTEXPR_AUTO from_string(const std::string &input) -> BigInt;
+
+class BigInt : public IBigIntLike
+{
+
+ public:
+	constexpr BigInt(): BigInt(0) {}
+
+	explicit constexpr BigInt(uint64_t v)
+		: _data({v})
+	{ }
+
+	explicit constexpr BigInt(const std::string &v)
+		: BigInt(std::move(from_string(v))) {}
+
+	explicit constexpr BigInt(const std::vector<uint64_t> &v)
+		: _data(v) {} // todo: maby make it non-public?
+
+	explicit constexpr BigInt(const std::vector<uint64_t> &&v)
+		: _data(std::move(v)) {} // todo: maby make it non-public?
+
+	constexpr ~BigInt() {}
+
+
+	constexpr BigInt(const BigInt &other) // copy constructor
+		: _data(other._data) {
+		LOGGING("copying: " << _data.size());
+	}
+
+
+	CONSTEXPR_AUTO_DISCARD operator=(const BigInt &other) -> BigInt& { // copy assignment
+		_data = other._data;
+		LOGGING("assigning: " << _data.size());
+		return *this;
+	}
+
+	// CONSTEXPR_AUTO_DISCARD operator=(const BigInt &&other) -> BigInt& { // move assignment
+	// 	todo: move assignement
+	// }
+
+	CONSTEXPR_AUTO size() const -> std::size_t {
+		return _data.size();
+	}
+
+	CONSTEXPR_AUTO operator[](std::size_t index) const -> uint64_t {
+		return (index >= size()) ? 0ull : _data.at(index);
+	}
+
+	CONSTEXPR_AUTO operator[](std::size_t index) -> uint64_t& {
+		if (index >= _data.size()) {
+			throw std::string(" index out of bound!");
+		}
+		return _data.at(index);
+	}
+
+	CONSTEXPR_VOID append(uint64_t v) { _data.push_back(v); }
+
+	CONSTEXPR_VOID insertFront(uint64_t v) { _data.insert(_data.begin(), v); }
+
+	CONSTEXPR_VOID cleanup() const { // todo: const???!
+		for (auto i = _data.size(); 0 <--i;) {
+			if (_data[i] == 0) {
+				_data.pop_back();
+			} else {
+				break;
+			}
+		}
+	}
+
+private:
+	CONSTEXPR_VOID resize(std::size_t size) {
+		_data.resize(size, 0);
+	}
+
+	friend CONSTEXPR_VOID resizeBigInt(BigInt &a, std::size_t size);
+
+public:
+	[[nodiscard]] auto  __data_for_testing_only() const -> const std::vector<uint64_t>&{
+		return _data;
+	}
+
+private:
+	mutable std::vector<uint64_t> _data; // only for cleanups! // todo: mutable???!
+	//uint32_t _lshift = 0;
+};
+
+
+CONSTEXPR_VOID resizeBigInt(BigInt &a, std::size_t size) {
+	a.resize(size);
+}
+
+CONSTEXPR_AUTO makeBigIntWithSize(std::size_t size) -> BigInt{
+	auto r = BigInt();
+	resizeBigInt(r, size); //OCTOPUS
+	return r;
+}
 
 template <typename T>
 struct BigIntBox_t {
 	using type = T;
 
 };
-
-class BigInt;
 
 template <>
 struct BigIntBox_t<BigInt> {
@@ -49,238 +158,81 @@ template <typename T>
 using BigIntBox = typename BigIntBox_t<T>::type;
 
 template <typename T>
-class BigIntUnaryOperation : IBigIntLike {
+class BigIntLShifted : IBigIntLike {
 public:
-	BigIntUnaryOperation(const BigIntBox<T> lhs) : _lhs(lhs) {}
+	constexpr BigIntLShifted(const BigIntBox<T> lhs, const uint64_t rhs) :
+		_lhs(lhs), _rhs(rhs) {}
 
-	inline std::size_t size() const { return _lhs.size(); }
-
-	inline uint64_t operator[](std::size_t index) const {
-		return _lhs[index];
+	CONSTEXPR_AUTO size() const -> std::size_t {
+		return lhs().size() + _rhs;
 	}
 
-	inline uint64_t& operator[](std::size_t index) {
-		return _lhs[index];
-	}
-protected:
-	const BigIntBox<T> _lhs;
-};
-
-template <typename TLHS, typename TRHS>
-class BigIntBinaryOperation : IBigIntLike {
-public:
-	BigIntBinaryOperation(const BigIntBox<TLHS> lhs, const BigIntBox<TRHS> rhs) :
-		_lhs(lhs),
-		_rhs(rhs) {}
-
-	inline std::size_t size() const { return _lhs.size(); }
-
-	inline uint64_t operator[](std::size_t index) const {
-		return _lhs[index];
+	CONSTEXPR_AUTO operator[](std::size_t index) const -> uint64_t {
+		return (index < _rhs) ? 0 : lhs()[index - _rhs];
 	}
 
-	inline uint64_t& operator[](std::size_t index) {
-		return _lhs[index];
-	}
-
-protected:
-	const BigIntBox<TLHS> lhs() {
-		return _lhs;
-	}
-
-	const BigIntBox<const TLHS> lhs() const {
-		return _lhs;
-	}
-
-	const BigIntBox<TRHS> rhs() {
-		return _rhs;
-	}
-
-	const BigIntBox<const TRHS> rhs() const {
-		return _rhs;
-	}
+//	inline uint64_t& operator[](std::size_t index) {
+//		return (index < rhs) ? 0 : lhs[index - _rhs]; // todo dont return 0 when out of bounds!
+//	}
 
 private:
-	const BigIntBox<TLHS> _lhs;
-	const BigIntBox<TRHS> _rhs;
-};
+	BigIntBox<T> _lhs;
+	uint64_t _rhs;
 
-
-inline static BigInt mult(uint64_t a, uint64_t b);
-
-std::string to_string_base2(const BigInt &v);
-std::string to_string_base8(const BigInt &v);
-std::string to_string_base10(const BigInt &v);
-std::string to_string_base16(const BigInt &v);
-std::string to_string(const BigInt &v);
-
-BigInt from_string_base2(const std::string &input);
-BigInt from_string_base8(const std::string &input);
-BigInt from_string_base10(const std::string &input);
-BigInt from_string_base16(const std::string &input);
-BigInt from_string(const std::string &input);
-
-class BigInt : IBigIntLike
-{
-
- public:
-	BigInt(): BigInt(0) {}
-
-	explicit BigInt(uint64_t v)
-	{ _data.push_back(v); }
-
-	explicit BigInt(const std::string &v)
-		: BigInt(std::move(from_string(v))) {}
-
-	explicit BigInt(const std::vector<uint64_t> &v)
-		: _data(v) {} // todo: maby make it non-public?
-
-	explicit BigInt(const std::vector<uint64_t> &&v)
-		: _data(std::move(v)) {} // todo: maby make it non-public?
-
-	BigInt(const std::initializer_list<uint64_t> v)
-		: _data(v) {} // todo: maby make it non-public?
-
-	~BigInt() {}
-
-
-	BigInt(const BigInt &other) :// copy constructor
-		_data(other._data) {
-		LOGGING("copying: " << _data.size());
-	}
-
-
-	inline BigInt &operator=(const BigInt &other) { // copy assignment
-		_data = other._data;
-		LOGGING("assigning: " << _data.size());
-		return *this;
-	}
-
-	inline std::size_t size() const { return _data.size(); }
-
-	inline uint64_t operator[](std::size_t index) const {
-		return (index >= size()) ? 0ull : _data.at(index);
-	}
-
-	inline uint64_t &operator[](std::size_t index)  {
-		if (index >= _data.size()) {
-			throw std::string(" index out of bound!");
-		}
-		return _data.at(index);
-	}
-
-	inline void append(uint64_t v) { _data.push_back(v); }
-
-	inline void insertFront(uint64_t v) { _data.insert(_data.begin(), v); }
-
-	inline void cleanup() const { // todo: const???!
-		for (auto i = _data.size(); 0 <--i;) {
-			if (_data[i] == 0) {
-				_data.pop_back();
-			} else {
-				break;
-			}
-		}
-	}
-
-private:
-	friend inline void resizeBigInt(BigInt &a, std::size_t size);
-
-public:
-	const std::vector<uint64_t>& __data_for_testing_only() const {
-		return _data;
-	}
-
-private:
-	mutable std::vector<uint64_t> _data; // only for cleanups! // todo: mutable???!
-	//uint32_t _lshift = 0;
-};
-
-
-template <typename TLHS>
-inline isBigIntLikeT1<TLHS, BigInt> toBigInt(const TLHS v) {
-	BigInt result;
-	for (int i = 0; i < v.size(); ++i) {
-		result.append(v[i]);
-	}
-	return result;
-}
-
-
-
-#define BIG_INT_BINARY_OPERATOR(name, tlhs, trhs) class name : BigIntBinaryOperation<tlhs, trhs> { \
-	using Base = BigIntBinaryOperation<tlhs, trhs>; \
-	public: \
-	name(BigIntBox<tlhs> lhs, BigIntBox<trhs> rhs) : Base(lhs, rhs) {}
-
-#define BIG_INT_UNARY_OPERATOR(name, tlhs) class name : BigIntUnaryOperation<tlhs> { \
-	using Base = BigIntUnaryOperation<tlhs>; \
-	name(BigIntBox<tlhs> lhs) : Base(lhs) {}
-
-
-
-inline void resizeBigInt(BigInt &a, std::size_t size) {
-	a._data.resize(size, 0);
-}
-
-inline BigInt makeBigIntWithSize(std::size_t size) {
-	auto r = BigInt();
-	resizeBigInt(r, size); //OCTOPUS
-	return r;
-}
-
-template <typename T>
-BIG_INT_BINARY_OPERATOR(BigIntLShifted, T, uint64_t)
-	inline std::size_t size() const { return Base::lhs().size() + Base::rhs(); }
-
-	inline uint64_t operator[](std::size_t index) const {
-		return (index < Base::rhs()) ? 0 : Base::lhs()[index - Base::rhs()];
-	}
-
-	inline uint64_t& operator[](std::size_t index) {
-		return (index < Base::rhs()) ? 0 : Base::lhs()[index - Base::rhs()]; // todo dont return 0 when out of bounds!
-	}
-
+	CONSTEXPR_AUTO lhs() const -> const T& { return _lhs; }
+	CONSTEXPR_AUTO lhs() -> T& { return _lhs; }
 };
 
 template <typename TLHS>
-inline auto lshifted(const TLHS &a, const uint64_t &b) {
+inline auto lshifted(const TLHS &a, uint64_t b) {
 	return BigIntLShifted<const TLHS>(a, b);
 }
 
-template <typename TLHS>
-inline auto lshiftedView(TLHS &a, const uint64_t &b) {
-	return BigIntLShifted<TLHS>(a, b);
-}
-
 template <typename T>
-BIG_INT_BINARY_OPERATOR(BigIntRShifted, T, uint64_t)
-	inline std::size_t size() const { return Base::lhs().size() - Base::rhs(); }
+class BigIntRShifted : IBigIntLike {
+public:
+	constexpr BigIntRShifted(const BigIntBox<T> lhs, const uint64_t rhs) :
+		_lhs(lhs), _rhs(rhs) {}
 
-	inline uint64_t operator[](std::size_t index) const {
+	CONSTEXPR_AUTO size() const -> std::size_t {
+		return lhs().size() - _rhs;
+	}
+
+	CONSTEXPR_AUTO operator[](std::size_t index) const -> uint64_t {
 		if (index < 0) {
 			throw std::string(" index out of bound!");
 		}
-//		Base::_lhs[index] = 7;
-		return Base::lhs()[index + Base::rhs()];
+		return lhs()[index + _rhs];
 	}
 
-	inline uint64_t &operator[](std::size_t index) {
+	CONSTEXPR_AUTO operator[](std::size_t index) -> uint64_t& {
 		if (index < 0) {
 			throw std::string(" index out of bound!");
 		}
-		return Base::lhs()[index + Base::rhs()];
+		return lhs()[index + _rhs];
 	}
 
+private:
+	BigIntBox<T> _lhs;
+	uint64_t _rhs;
+
+	CONSTEXPR_AUTO lhs() const -> const T& { return _lhs; }
+	CONSTEXPR_AUTO lhs() -> T& { return _lhs; }
 };
 
 template <typename TLHS>
-inline auto rshifted(TLHS &a, const uint64_t &b) {
+CONSTEXPR_AUTO rshifted(const TLHS &a, uint64_t b) {
+	return BigIntRShifted<const TLHS>(a, b);
+}
+
+template <typename TLHS>
+CONSTEXPR_AUTO rshifted(TLHS &a, uint64_t b) {
 	return BigIntRShifted<TLHS>(a, b);
 }
 
 template <typename TLHS, typename TRHS>
-inline isBigIntLikeT2<TLHS, TRHS, bool> operator<(const TLHS &a, const TRHS &b) {
+	requires is_BigInt_like2<TLHS, TRHS>
+CONSTEXPR_AUTO operator<(const TLHS &a, const TRHS &b) {
 	std::size_t i;
 	for (i = std::max(a.size(), b.size()); i-->1 && a[i] == b[i];) {
 		// do nothin'
@@ -289,7 +241,8 @@ inline isBigIntLikeT2<TLHS, TRHS, bool> operator<(const TLHS &a, const TRHS &b) 
 }
 
 template <typename TLHS, typename TRHS>
-inline isBigIntLikeT2<TLHS, TRHS, bool> operator>(const TLHS &a, const TRHS &b) {
+	requires is_BigInt_like2<TLHS, TRHS>
+CONSTEXPR_AUTO operator>(const TLHS &a, const TRHS &b) {
 	std::size_t i;
 	for (i = std::max(a.size(), b.size()); i-->1 && a[i] == b[i];) {
 		// do nothin'
@@ -298,18 +251,21 @@ inline isBigIntLikeT2<TLHS, TRHS, bool> operator>(const TLHS &a, const TRHS &b) 
 }
 
 template <typename TLHS>
-inline isBigIntLikeT1<TLHS, bool> operator<(const TLHS &a, const uint64_t &b) {
+	requires is_BigInt_like1<TLHS>
+CONSTEXPR_AUTO operator<(const TLHS &a, const uint64_t &b) {
 	return a < BigInt(b);
 }
 
 template <typename TLHS>
-inline isBigIntLikeT1<TLHS, bool> operator>(const TLHS &a, const uint64_t &b) {
+	requires is_BigInt_like1<TLHS>
+CONSTEXPR_AUTO operator>(const TLHS &a, const uint64_t &b) {
 	return a > BigInt(b);
 }
 
 
 template <typename TRES, typename TLHS, typename TRHS>
-inline static isBigIntLikeT2<TRES, TLHS, void> add(TRES &result, TLHS &a, const TRHS &b) {
+	requires is_BigInt_like3<TRES, TLHS, TRHS>
+CONSTEXPR_VOID add(TRES &result, TLHS &a, const TRHS &b) {
 	uint8_t c = 0; // carry
 	for (auto i = 0ull; i < result.size(); i++) {
 		const auto bc = b[i] + c;
@@ -323,14 +279,16 @@ inline static isBigIntLikeT2<TRES, TLHS, void> add(TRES &result, TLHS &a, const 
 }
 
 template <typename TLHS, typename TRHS>
-inline isBigIntLikeT2<TLHS, TRHS, BigInt> operator+(const TLHS &a, const TRHS &b) {
+	requires is_BigInt_like2<TLHS, TRHS>
+CONSTEXPR_AUTO operator+(const TLHS &a, const TRHS &b) -> BigInt {
 	BigInt result = makeBigIntWithSize(std::max(a.size(), b.size()));
 	add(result, a, b);
 	return result;
 }
 
-template <typename TRHS>
-inline isBigIntLikeT1<TRHS, BigInt>& operator+=(BigInt &a, const TRHS &b) {
+template <typename TLHS, typename TRHS>
+	requires is_BigInt_like2<TLHS, TRHS>
+CONSTEXPR_AUTO_DISCARD operator+=(TLHS &a, const TRHS &b) -> TLHS& {
 	resizeBigInt(a, std::max(a.size(), b.size()));
 	add(a, a, b);
 	return a;
@@ -338,7 +296,8 @@ inline isBigIntLikeT1<TRHS, BigInt>& operator+=(BigInt &a, const TRHS &b) {
 
 
 template <typename TRES, typename TLHS, typename TRHS>
-inline static isBigIntLikeT2<TRES, TLHS, void> sub(TRES &result, TLHS &a, const TRHS &b) {
+	requires is_BigInt_like3<TRES, TLHS, TRHS>
+CONSTEXPR_VOID sub(TRES &result, TLHS &a, const TRHS &b) {
 	uint8_t c = 0; // carry
 	for (auto i = 0ull; i < result.size(); i++) {
 		const auto bc = b[i] + c;
@@ -352,7 +311,8 @@ inline static isBigIntLikeT2<TRES, TLHS, void> sub(TRES &result, TLHS &a, const 
 }
 
 template <typename TLHS, typename TRHS>
-inline isBigIntLikeT2<TLHS, TRHS, BigInt> operator-(const TLHS &a, const TRHS &b) {
+	requires is_BigInt_like2<TLHS, TRHS>
+CONSTEXPR_AUTO operator-(const TLHS &a, const TRHS &b) -> BigInt {
 	BigInt result = makeBigIntWithSize(std::max(a.size(), b.size()));
 	sub(result, a, b);
 	result.cleanup();
@@ -360,13 +320,14 @@ inline isBigIntLikeT2<TLHS, TRHS, BigInt> operator-(const TLHS &a, const TRHS &b
 }
 
 template <typename TLHS, typename TRHS>
-inline isBigIntLikeT2<TLHS, TRHS, TLHS>& operator-=(TLHS &a, const TRHS &b) {
+	requires is_BigInt_like2<TLHS, TRHS>
+CONSTEXPR_AUTO_DISCARD operator-=(TLHS &a, const TRHS &b) -> TLHS& {
 	sub(a, a, b);
 	return a;
 }
 
 
-inline static BigInt mult(uint64_t a, uint64_t b) {
+CONSTEXPR_AUTO mult(uint64_t a, uint64_t b) -> BigInt {
 	const auto a_0 = a & 0xFFFFFFFFull;
 	const auto a_1 = a >> 32;
 	const auto b_0 = b & 0xFFFFFFFFull;
@@ -385,11 +346,12 @@ inline static BigInt mult(uint64_t a, uint64_t b) {
 	c += (r < (r_01 << 32)) ? 1 : 0;
 	r += r_10 << 32;
 	c += (r < (r_10 << 32)) ? 1 : 0;
-	return {r, c};
+	return BigInt{std::vector<uint64_t>{r, c}};
 }
 
 template <typename TRES, typename TLHS>
-inline static isBigIntLikeT2<TRES, TLHS, void> mult(TRES &result, TLHS &a, uint64_t b) {
+	requires is_BigInt_like2<TRES, TLHS>
+CONSTEXPR_VOID mult(TRES &result, TLHS &a, uint64_t b) {
 	uint64_t c = 0; // carry
 	for (auto i = 0ull; i < a.size(); i++) {
 		const auto rc = mult(a[i], b);
@@ -402,36 +364,44 @@ inline static isBigIntLikeT2<TRES, TLHS, void> mult(TRES &result, TLHS &a, uint6
 	LOGGING("multiplying: " << result.size());
 }
 
-template <typename TLHS>
-inline isBigIntLikeT1<TLHS, BigInt> operator*(const TLHS &a, uint64_t b) {
-	BigInt result = makeBigIntWithSize(a.size());
-	mult(result, a, b);
-	return  result;
+template <typename TRES, typename TLHS, typename TRHS>
+	requires is_BigInt_like3<TRES, TLHS, TRHS>
+CONSTEXPR_VOID mult(TRES &result, const TLHS &a, const TRHS &b) {
+	for (auto i = 0ull; i < b.size(); i++) {
+		auto temp = a * b[i];
+		result += lshifted(temp, i);
+	}
+	LOGGING("multiplying: " << result.size());
 }
 
 template <typename TLHS>
-inline isBigIntLikeT1<TLHS, BigInt> &operator*=(TLHS &a, uint64_t b) {
+	requires is_BigInt_like1<TLHS>
+CONSTEXPR_AUTO operator*(const TLHS &a, uint64_t b) -> BigInt {
+	BigInt result = makeBigIntWithSize(a.size());
+	mult(result, a, b);
+	return result;
+}
+
+template <typename TLHS>
+	requires is_BigInt_like1<TLHS>
+CONSTEXPR_AUTO_DISCARD operator*=(TLHS &a, uint64_t b) -> BigInt& {
 	mult(a, a, b);
 	return a;
 }
 
 template <typename TLHS, typename TRHS>
-inline isBigIntLikeT2<TLHS, TRHS, BigInt> operator*(const TLHS &a, const TRHS &b) {
+	requires is_BigInt_like2<TLHS, TRHS>
+CONSTEXPR_AUTO operator*(const TLHS &a, const TRHS &b) -> BigInt {
 	BigInt result = makeBigIntWithSize(a.size() + b.size());
-
-	for (auto i = 0ull; i < b.size(); i++) {
-		auto temp = a * b[i];
-		result += lshifted(temp, i);
-	}
-
+	mult(result, a, b);
 	result.cleanup();
-	LOGGING("multiplying: " << result.size());
 	return result;
 }
 
 
 template <typename TLHS>
-inline isBigIntLikeT1<TLHS, BigInt> lshiftBits(const TLHS &vec, uint64_t digits) {
+	requires is_BigInt_like1<TLHS>
+CONSTEXPR_AUTO lshiftBits(const TLHS &vec, uint64_t digits) -> BigInt {
 	BigInt result(std::vector<uint64_t>(digits / 64, 0));
 	digits %= 64;
 	result.append(vec[0] << digits);
@@ -447,7 +417,8 @@ inline isBigIntLikeT1<TLHS, BigInt> lshiftBits(const TLHS &vec, uint64_t digits)
 }
 
 template <typename TLHS>
-inline isBigIntLikeT1<TLHS, BigInt> rshiftBits(const TLHS &vec, uint64_t digits) {
+	requires is_BigInt_like1<TLHS>
+CONSTEXPR_AUTO rshiftBits(const TLHS &vec, uint64_t digits) -> BigInt {
 	BigInt result;//(digits / 64, 0);
 	const uint64_t start = digits / 64;
 	digits %= 64;
@@ -460,8 +431,9 @@ inline isBigIntLikeT1<TLHS, BigInt> rshiftBits(const TLHS &vec, uint64_t digits)
 	return result;
 }
 
-template <typename TRHS>
-inline isBigIntLikeT1<TRHS, DivModResult<BigInt>> divmod(const BigInt &a, const TRHS &b) {
+template <typename TLHS, typename TRHS>
+	requires is_BigInt_like2<TLHS, TRHS>
+CONSTEXPR_AUTO divmod(const TLHS &a, const TRHS &b) -> DivModResult<BigInt> {
 	DivModResult<BigInt> r;
 	r.r = a;
 	resizeBigInt(r.d, a.size() - b.size() + 1);
@@ -477,8 +449,7 @@ inline isBigIntLikeT1<TRHS, DivModResult<BigInt>> divmod(const BigInt &a, const 
 	const auto b2d = b.size()*64 - __builtin_clzll(b2Hi);
 	const auto adbd2 = int64_t(ad - b2d);
 	//uint64_t min = std::max(0ll, adbd2 - 1);
-	uint64_t max = std::max(0ll, adbd2 + 1ll);//%64; //std::numeric_limits<uint64_t>::max();
-	uint64_t k;
+	const uint64_t max = std::max(0ll, adbd2 + 1ll);//%64; //std::numeric_limits<uint64_t>::max();
 	uint64_t i = max;//(a.size() - b.size())*64;
 	auto a2IsSmaller = false;
 
@@ -497,13 +468,16 @@ inline isBigIntLikeT1<TRHS, DivModResult<BigInt>> divmod(const BigInt &a, const 
 	return r;
 }
 
-inline DivModResult<BigInt, uint64_t> divmod(const BigInt &a, const uint64_t &b) {
+template <typename TLHS>
+	requires is_BigInt_like1<TLHS>
+CONSTEXPR_AUTO divmod(const TLHS &a, const uint64_t &b) -> DivModResult<BigInt, uint64_t> {
 	const auto res = divmod(a, BigInt(b));
 	return { res.d, res.r[0] };
 }
 
 template <typename TLHS>
-inline DivModResult<BigInt, uint64_t> divmod1(const TLHS &a, const uint32_t &b) {
+	requires is_BigInt_like1<TLHS>
+CONSTEXPR_AUTO divmod1(const TLHS &a, const uint32_t &b) -> DivModResult<BigInt, uint64_t> { // todo: validate: Isn't uint32_t sufficient for remainder return type?
 	BigInt d;
 	resizeBigInt(d, a.size());
 	BigInt y;
@@ -548,7 +522,8 @@ inline DivModResult<BigInt, uint64_t> divmod1(const TLHS &a, const uint32_t &b) 
 // Division:
 
 template <typename TLHS>
-inline isBigIntLikeT1<TLHS, BigInt> operator/(const TLHS &a, const uint32_t &b) {
+	requires is_BigInt_like1<TLHS>
+CONSTEXPR_AUTO operator/(const TLHS &a, const uint32_t &b) -> BigInt {
 	BigInt result;
 	resizeBigInt(result, a.size());
 
@@ -570,19 +545,22 @@ inline isBigIntLikeT1<TLHS, BigInt> operator/(const TLHS &a, const uint32_t &b) 
 }
 
 template <typename TLHS>
-inline isBigIntLikeT1<TLHS, BigInt> operator/(const TLHS &a, const uint64_t &b) {
+	requires is_BigInt_like1<TLHS>
+CONSTEXPR_AUTO operator/(const TLHS &a, const uint64_t &b) -> BigInt {
 	return divmod(a, b).d;
 }
 
 template <typename TLHS, typename TRHS>
-inline isBigIntLikeT2<TLHS, TRHS, BigInt> operator/(const TLHS &a, const TRHS &b) {
+	requires is_BigInt_like2<TLHS, TRHS>
+CONSTEXPR_AUTO operator/(const TLHS &a, const TRHS &b) -> BigInt {
 	return divmod(a, b).d;
 }
 
 // Modulo:
 
 template <typename TLHS>
-inline isBigIntLikeT1<TLHS, uint32_t> operator%(const TLHS &a, const uint32_t &b) {
+	requires is_BigInt_like1<TLHS>
+CONSTEXPR_AUTO operator%(const TLHS &a, const uint32_t &b) -> uint32_t {
 	BigInt y;
 	resizeBigInt(y, a.size());
 
@@ -620,19 +598,152 @@ inline isBigIntLikeT1<TLHS, uint32_t> operator%(const TLHS &a, const uint32_t &b
 	return r[0];
 }
 
-inline uint64_t operator%(const BigInt &a, const uint64_t &b) {
+template <typename TLHS>
+	requires is_BigInt_like1<TLHS>
+CONSTEXPR_AUTO operator%(const TLHS &a, const uint64_t &b) -> uint64_t {
 	return divmod(a, b).r;
 }
 
 template <typename TLHS, typename TRHS>
-inline isBigIntLikeT2<TLHS, TRHS, BigInt> operator%(const TLHS &a, const TRHS &b) {
+	requires is_BigInt_like2<TLHS, TRHS>
+CONSTEXPR_AUTO operator%(const TLHS &a, const TRHS &b) -> BigInt {
 	return divmod(a, b).r;
 }
 
 
-BigInt factorial(uint32_t n);
+namespace _private {
+struct base_conversion {
+	constexpr base_conversion(uint32_t base) noexcept
+		: base_power(floor(64 / log2(base))), //  = 19 for base 10;
+		division_base(std::pow(base, base_power)) {}
+	uint32_t base_power;
+	uint64_t division_base;
+};
 
-uint64_t toDigitSum(const BigInt &v);
+template <int base = 10>
+inline std::string to_string_padded_generic(uint64_t n, uint64_t len){
+	std::string result(len, '0');
+	for (uint64_t val = n; len-->0 && val != 0; val/=base)
+		result[len] = '0' + val%base;
+	return result;
+}
+
+template <int base>
+CONSTEXPR_AUTO to_string_generic(const BigInt &v) -> std::string {
+	const auto conv = base_conversion(base);
+
+	DivModResult<BigInt, uint64_t> tempDig = { v, 0ull };
+	tempDig.d.cleanup();
+
+	std::string result;
+	while (tempDig.d > 0) {
+		tempDig = divmod(tempDig.d, conv.division_base);
+		auto digs = tempDig.r;
+		result.insert(0, to_string_padded_generic<base>(digs, conv.base_power));
+		tempDig.d.cleanup();
+	}
+	//       15511210043330985984000000
+	// 25! = 1551121  43330985984000000
+	const auto index = result.find_first_of("123456789ABCDEF");
+	result.erase(0, index);
+	if (result.size() == 0) {
+		result = "0";
+	}
+	return result;
+}
+
+template <int base>
+CONSTEXPR_AUTO from_string_generic(const std::string &input) -> BigInt {
+	const auto conv = base_conversion(base);
+
+	BigInt result{0};
+
+	for (size_t i = 0; i < input.size(); i += conv.base_power) {
+		auto substr = std::string_view(input).substr(i, conv.base_power);
+		auto mul = substr.size() == conv.base_power ? conv.division_base : uint64_t(std::pow(base, substr.size()));
+		auto add = utils::stoull(substr, base);
+		result *= mul;
+		result += BigInt{add};
+	}
+	//	result.cleanup();
+	return result;
+}
+
+}
+
+CONSTEXPR_AUTO factorial(uint32_t n) -> BigInt {
+	BigInt result(1);
+	for (uint32_t i = 1; i <= n; ++i) {
+		result *= i;
+	}
+	return result;
+}
+
+CONSTEXPR_AUTO toDigitSum(const BigInt &v) -> uint64_t {
+	// 4294967296 =>
+	// 1000000000
+	// 1111111110
+	// 18446744073709551616 =>
+	// 10000000000000000000
+	const auto base = _private::base_conversion{10}.division_base; // 19 is the larges value for n such that 10^n fits into 64 bits
+	// constexpr auto base = uint64_t(10000000000000000000ull);
+	DivModResult<BigInt, uint64_t> tempDig = { v, 0ull };
+	tempDig.d.cleanup();
+
+	uint64_t sum = 0ull;
+	while (tempDig.d > 0) {
+		tempDig = divmod(tempDig.d, base);
+		auto digs = tempDig.r;
+		while (digs > 00) {
+			sum += digs % 10;
+			digs /= 10;
+		}
+		tempDig.d.cleanup();
+	}
+	return sum;
+}
+
+CONSTEXPR_AUTO to_string_base2(const BigInt &v) -> std::string {
+	return _private::to_string_generic<2>(v);
+}
+
+CONSTEXPR_AUTO to_string_base8(const BigInt &v) -> std::string {
+	return _private::to_string_generic<8>(v);
+}
+
+CONSTEXPR_AUTO to_string_base10(const BigInt &v) -> std::string {
+	return _private::to_string_generic<10>(v);
+}
+
+CONSTEXPR_AUTO to_string_base16(const BigInt &v) -> std::string {
+	return _private::to_string_generic<16>(v);
+}
+
+CONSTEXPR_AUTO to_string(const BigInt &v) -> std::string {
+	return to_string_base10(v);
+}
+
+
+CONSTEXPR_AUTO from_string_base2(const std::string &input) -> BigInt {
+	return _private::from_string_generic<2>(input);
+}
+
+CONSTEXPR_AUTO from_string_base8(const std::string &input) -> BigInt {
+	return _private::from_string_generic<8>(input);
+}
+
+CONSTEXPR_AUTO from_string_base10(const std::string &input) -> BigInt {
+	return _private::from_string_generic<10>(input);
+}
+
+CONSTEXPR_AUTO from_string_base16(const std::string &input) -> BigInt {
+	return _private::from_string_generic<16>(input);
+}
+
+CONSTEXPR_AUTO from_string(const std::string &input) -> BigInt {
+	return from_string_base10(input);
+}
+
 
 inline std::ostream & operator<<(std::ostream &ostream, const BigInt &val) {
 	return ostream << to_string(val);
